Game Store Shopping Cart and Orders Implementation
This project is part of a game store application built with Java and Spring Boot. The focus of this exercise is the implementation of a shopping cart service and order management system that allows users to:

Add games to their shopping cart.
Remove games from the shopping cart.
Place orders to purchase games, which records details about the user and the purchased games into the database.